-simplify tag handling by using list-based tags instead of custom Tag type: - Replace Tag constructors (ETag, PTag, etc) with list-based tags per NIP-01 - Move types into dedicated modules (Event, Profile, Relay) - Update tag parsing/generation throughout codebase - Simplify root/reply reference handling in UI
