+++ src/RelayMgmt.hs
@@ -13,8 +13,8 @@ import Nostr
 import Nostr.Event (createPreferredDMRelaysEvent, createRelayListMetadataEvent)
 import Nostr.Keys (PubKeyXO, keyPairToPubKeyXO)
 import Nostr.Publisher
+import Nostr.Relay (Relay(..), RelayURI, defaultDMRelays, defaultGeneralRelays, getUri)
 import Nostr.RelayConnection
-import Nostr.Types (Relay(..), RelayURI, defaultDMRelays, defaultGeneralRelays, getUri)
 import Nostr.Util
 import Store.Lmdb (LmdbStore, getDMRelays, getGeneralRelays)
